example/python-vulnerable-app/requirements.txt
flask==3.0.0
werkzeug==3.1.3
requests==2.31.0
PyJWT==2.8.0
PyYAML==6.0.1
Jinja2==3.1.2
